fix vgg

+++ dataset/InpaintDataset.py
@@ -73,11 +73,15 @@ class SCDataset(Dataset):
         colormap = (cv2.imread(pathes["colormap"], -1) / 255 - 0.5) * 2
         sketch = cv2.imread(pathes["sketch"], -1)
 
+        image = image.astype(np.float16)
+        colormap = colormap.astype(np.float16)
+        sketch = sketch.astype(np.float16)
+
         assert image is not None
         assert colormap is not None, f"{pathes['colormap']} does't exist"
         assert sketch is not None
         if self.sc_only:
-            mask = np.zeros_like(image[:, :, 0], dtype=np.float32)
+            mask = np.zeros_like(image[:, :, 0])
         else:
             mask = self.user_simulator(image)
